+++ src/Scene/objects/Icosahedron.java
@@ -57,8 +57,8 @@ public class Icosahedron extends Shape {
     }
 
     public Icosahedron(float x, float y, float z, JSONObject size, Color color) {
-        super(x, y, z, (float) size.get("scale"), color);
+        super(x, y, z, ((Double) size.get("scale")).floatValue(), color);
 
-        generateVertices((float) size.get("width"), (float) size.get("length"), (float) size.get("height"));
+        generateVertices(((Double) size.get("width")).floatValue(), ((Double) size.get("length")).floatValue(), ((Double) size.get("height")).floatValue());
     }
 }
