+++ src/Scene/objects/Plane.java
@@ -36,8 +36,8 @@ public class Plane extends Shape {
     }
 
     public Plane(float x, float y, float z, JSONObject size, Color color) {
-        super(x, y, z, (float) size.get("scale"), color);
+        super(x, y, z, ((Double) size.get("scale")).floatValue(), color);
 
-        generateVertices((float) size.get("width"), (float) size.get("length"));
+        generateVertices(((Double) size.get("width")).floatValue(), ((Double) size.get("length")).floatValue());
     }
 }
