+++ src/Scene/objects/Shape.java
@@ -30,9 +30,9 @@ public class Shape {
         Object colorObj = object.get("color");
 
         // coordinate handling
-        float x = (float) coordinates.get(0);
-        float y = (float) coordinates.get(1);
-        float z = (float) coordinates.get(2);
+        float x = ((Double) coordinates.get(0)).floatValue();
+        float y = ((Double) coordinates.get(1)).floatValue();
+        float z = ((Double) coordinates.get(2)).floatValue();
 
         // color handling
         Color color;
